requirements.txt
Django~=5.0.6
django-environ==0.11.2
environ~=1.0
djangorestframework~=3.15.1
django-extensions>=3.1.3
django-filter==2.4.0
